2024/CVE-2024-47177/requirements.txt
pycups==2.0.4
